# blockchain-api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: KALDRIX Blockchain API
  description: REST and WebSocket API for KALDRIX blockchain operations including DAG management, quantum-resistant cryptography, and PBFT consensus
  version: 1.0.0
  contact:
    name: KALDRIX Team
    email: api@kaldrix.io
    url: https://kaldrix.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.kaldrix.io
    description: Production server
paths:
...
  /api/v1/wallet/create:
    post:
      summary: Create Wallet
      description: Create a new wallet
      tags:
        - Wallet
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: My Wallet
                password:
                  type: string
                  format: password
                type:
                  type: string
                  enum: [standard, hardware, multisig]
                  default: standard
      responses:
        '201':
          description: Wallet created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  walletId:
                    type: string
                  address:
                    type: string
                  publicKey:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  warning:
                    type: string
  /api/v1/wallet/{walletId}:
    get:
      summary: Get Wallet
      description: Retrieve wallet information
      tags:
        - Wallet
      security:
        - bearerAuth: []
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wallet retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  walletId:
                    type: string
                  address:
                    type: string
                  publicKey:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
                  balance:
                    type: number
                  createdAt:
                    type: string
                    format: date-time
                  lastActivity:
                    type: string
                    format: date-time
                  transactionCount:
                    type: integer
                  isEncrypted:
                    type: boolean
        '404':
          description: Wallet not found
  /api/v1/wallet/{walletId}/balance:
    get:
      summary: Get Wallet Balance
      description: Retrieve wallet balance information
      tags:
        - Wallet
      security:
        - bearerAuth: []
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  walletId:
                    type: string
                  address:
                    type: string
                  balance:
                    type: number
                  pendingBalance:
                    type: number
                  lastUpdated:
                    type: string
                    format: date-time

  # Consensus Operations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - bearerAuth: []
  - apiKeyAuth: []